refactor(cli): move network options from global options
