+++ packages/tools/kadena-cli/src/utils/globalOptions.ts
@@ -27,7 +27,7 @@ import {
 import { readKeyFileContent } from '../keys/utils/storage.js';
 import { loadNetworkConfig } from '../networks/utils/networkHelpers.js';
 import { createOption } from './createOption.js';
-import { isNotEmptyString, passwordPromptTransform } from './helpers.js';
+import { passwordPromptTransform } from './helpers.js';
 import { log } from './logger.js';
 
 // eslint-disable-next-line @rushstack/typedef-var
@@ -109,84 +109,6 @@ export const globalOptions = {
     ),
   }),
   // Network
-  networkName: createOption({
-    key: 'networkName' as const,
-    prompt: networks.networkNamePrompt,
-    validation: z.string(),
-    option: new Option(
-      '-n, --network-name <networkName>',
-      'Kadena network (e.g. "mainnet")',
-    ),
-    transform: (networkName: string) => {
-      const trimmedNetworkName = networkName.trim();
-      if (isNotEmptyString(trimmedNetworkName)) {
-        return trimmedNetworkName;
-      }
-
-      throw new Error('Network name is required');
-    },
-  }),
-  networkId: createOption({
-    key: 'networkId' as const,
-    prompt: networks.networkIdPrompt,
-    validation: z.string(),
-    option: new Option(
-      '--network-id <networkId>',
-      'Kadena network Id (e.g. "mainnet01")',
-    ),
-    transform: (networkId: string) => {
-      return networkId.trim();
-    },
-  }),
-  networkHost: createOption({
-    key: 'networkHost' as const,
-    prompt: networks.networkHostPrompt,
-    validation: z.string(),
-    option: new Option(
-      '-h, --network-host <networkHost>',
-      'Kadena network host (e.g. "https://api.chainweb.com")',
-    ),
-    transform: (value: string) => {
-      // when it's optional and it's empty string and we don't want to validate it
-      if (isNotEmptyString(value)) {
-        const parse = z.string().url().safeParse(value.trim());
-        if (!parse.success) {
-          throw new Error(
-            'Network host: Invalid URL. Please enter a valid URL.',
-          );
-        }
-      }
-
-      return value.trim();
-    },
-  }),
-  networkExplorerUrl: createOption({
-    key: 'networkExplorerUrl' as const,
-    prompt: networks.networkExplorerUrlPrompt,
-    validation: z.string().optional(),
-    option: new Option(
-      '-e, --network-explorer-url <networkExplorerUrl>',
-      'Kadena network explorer URL (e.g. "https://explorer.chainweb.com/mainnet/tx/")',
-    ),
-  }),
-  networkOverwrite: createOption({
-    key: 'networkOverwrite' as const,
-    prompt: networks.networkOverwritePrompt,
-    validation: z.string(),
-    option: new Option(
-      '-o, --network-overwrite <networkOverwrite>',
-      'Overwrite existing network configuration (yes/no)',
-    ),
-  }),
-  networkDelete: createOption({
-    key: 'networkDelete' as const,
-    prompt: networks.networkDeletePrompt,
-    validation: z.string(),
-    option: new Option(
-      '-d, --network-delete <networkDelete>',
-      'Delete the configuration for network (yes/no)',
-    ),
-  }),
   network: createOption({
     key: 'network' as const,
     prompt: networks.networkSelectPrompt,
